Added BrickBreaker Game
The game is a basic brickbreaker game with Basic Graphics and layout. The game consists of a paddle to with which we can navigate a ball after impact with it. The ball has to hit the bricks to increment the score by 5. After hitting the bricks, the bricks disintegrate or get destroyed and the ball can now move through the space of the Play area.